feat(notification): integrate Redis for real-time notifications and enhance notification handling
- Added Redis adapter for Socket.IO in the gateway to enable real-time notifications.
- Updated AppNotification entity to include clientType and notification type.
- Refactored AppService to handle notification creation and sending more efficiently.
- Implemented AppNotificationWorker to process notification jobs.
- Enhanced notification DTOs to include new fields and validation.
- Created NotificationGateway and NotificationService for handling WebSocket connections and events.
- Added RabbitMQ integration for notification events.
- Updated database initialization script to include notification_db.
- Added tests for NotificationGateway and NotificationService.

+++ apps/notification/src/app/app.service.ts
@@ -1,21 +1,35 @@
-import { BadRequestException, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  Inject,
+  Injectable,
+  InternalServerErrorException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { AppNotification } from './entitities/app-notifications.entity';
 import { Repository } from 'typeorm';
 import { CreateAppNotificationDto } from './dto/create-app-notification.dto';
 import { GetUserNotificationsDto } from './dto/get-user-notifications.dto';
+import { RABBITMQ_QUEUES } from '@app/shared/utils/constants';
+import { ClientProxy } from '@nestjs/microservices';
+import { lastValueFrom } from 'rxjs';
 
 @Injectable()
 export class AppService {
   constructor(
     @InjectRepository(AppNotification)
     private appNotificationRepository: Repository<AppNotification>,
+    @Inject(RABBITMQ_QUEUES.GATEWAY) private gatewayProxy: ClientProxy,
   ) {}
 
-  async sendNotification(createAppNotificationDto: CreateAppNotificationDto) {
+  async createNotification(createAppNotificationDto: CreateAppNotificationDto) {
     const notification = await this.appNotificationRepository.save(
       createAppNotificationDto,
     );
+    const notifications = await this.getUserNotifications({
+      userId: createAppNotificationDto.userId,
+    });
+    this.sendNotificationsToUser(notifications);
+    if (createAppNotificationDto.popup) this.sendPopupToUser(notification);
     return notification;
   }
 
@@ -49,7 +63,12 @@ export class AppService {
       throw new BadRequestException('Notification not found');
     }
     notification.isRead = true;
-    return notification.save();
+    await notification.save();
+    const notifications = await this.getUserNotifications({
+      userId: notification.userId,
+    });
+    this.sendNotificationsToUser(notifications);
+    return notification;
   }
 
   async markAllAsRead(userId: string) {
@@ -57,16 +76,91 @@ export class AppService {
       { userId, isRead: false },
       { isRead: true },
     );
+    const notifications = await this.getUserNotifications({ userId });
+    this.sendNotificationsToUser(notifications);
   }
 
   async deleteNotification(notificationId: string) {
-    const result = await this.appNotificationRepository.delete(notificationId);
-    if (result.affected === 0) {
+    const notification = await this.appNotificationRepository.findOneBy({
+      id: notificationId,
+    });
+    if (!notification) {
       throw new BadRequestException('Notification not found');
     }
+    await notification.remove();
+    const notifications = await this.getUserNotifications({
+      userId: notification.userId,
+    });
+    this.sendNotificationsToUser(notifications);
+    return notification;
   }
 
   async deleteAllNotifications(userId: string) {
+    const notification = await this.appNotificationRepository.findOneBy({
+      userId,
+    });
+    if (!notification) {
+      throw new BadRequestException('No notifications found for user');
+    }
     await this.appNotificationRepository.delete({ userId });
+    if (notification.clientType === 'vendor') {
+      this.gatewayProxy.emit('vendor.notifyUser', {
+        userId: notification.userId,
+        notifications: [],
+      });
+    } else if (notification.clientType === 'client') {
+      this.gatewayProxy.emit('client.notifyUser', {
+        userId: notification.userId,
+        notifications: [],
+      });
+    }
+  }
+
+  sendNotificationsToUser(notifications: AppNotification[]) {
+    if (notifications[0].clientType === 'vendor') {
+      lastValueFrom(
+        this.gatewayProxy.emit('vendor.notifyUser', {
+          userId: notifications[0].userId,
+          notifications,
+        }),
+      ).catch((err) => {
+        console.error(err);
+        throw new InternalServerErrorException('Failed to notify vendor');
+      });
+    } else if (notifications[0].clientType === 'client') {
+      lastValueFrom(
+        this.gatewayProxy.emit('client.notifyUser', {
+          userId: notifications[0].userId,
+          notifications,
+        }),
+      ).catch((err) => {
+        console.error(err);
+        throw new InternalServerErrorException('Failed to notify client');
+      });
+    }
+  }
+
+  sendPopupToUser(notification: AppNotification) {
+    if (notification.clientType === 'vendor') {
+      lastValueFrom(
+        this.gatewayProxy.emit('vendor.popupNotification', {
+          userId: notification.userId,
+          notification,
+        }),
+      ).catch((err) => {
+        console.error(err);
+        throw new InternalServerErrorException('Failed to popup vendor');
+      });
+    } else if (notification.clientType === 'client') {
+      lastValueFrom(
+        this.gatewayProxy.emit('client.popupNotification', {
+          userId: notification.userId,
+          notification,
+        }),
+      ).catch((err) => {
+        console.error(err);
+        throw new InternalServerErrorException('Failed to popup client');
+      });
+    }
   }
 }
